requirements.txt
google-generativeai
discord.py
python-dotenv
google-genai